more correction of types
git-svn-id: https://svn.r-project.org/R/trunk@42242 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/gnuwin32/console.c
@@ -162,7 +162,7 @@ void xbufgrow(xbuf p, xlong dim, xint ms)
     if(dim > p->dim) {
 	ret = realloc(p->b, dim + 1);
 	if(ret) {
-	    change = (unsigned long) ret - (unsigned long) p->b;
+	    change = (uintptr_t) ret - (uintptr_t) p->b;
 	    p->b = (char *) ret;
 	    p->av += change;
 	    p->free += change;
